+++ core/src/main/java/jvmram/process/JvmProcessInfo.java
@@ -2,10 +2,10 @@ package jvmram.process;
 
 import org.jspecify.annotations.NonNull;
 
-public record JvmProcessInfo(long pid, String displayName) implements Comparable<JvmProcessInfo> {
+public record JvmProcessInfo(int pid, String displayName) implements Comparable<JvmProcessInfo> {
     @Override
     public int compareTo(@NonNull JvmProcessInfo other) {
-        return Long.compare(pid, other.pid);
+        return Integer.compare(pid, other.pid);
     }
 }
