+++ core/src/main/java/jvmram/suppliers/HardwareDataSuppliersFactory.java
@@ -4,7 +4,7 @@ import jvmram.model.metrics.MetricType;
 import jvmram.suppliers.data.HardwareData;
 
 public interface HardwareDataSuppliersFactory {
-    <T extends HardwareData> HardwareDataSupplier<T> getOrCreateSupplier(long pid, MetricType metricType);
+    <T extends HardwareData> HardwareDataSupplier<T> getOrCreateSupplier(int pid, MetricType metricType);
 
     static HardwareDataSuppliersFactory getInstance() {
         return HardwareDataSuppliersFactoryImpl.INSTANCE;
