+++ model/src/main/java/jvmram/model/graph/GraphKey.java
@@ -2,5 +2,5 @@ package jvmram.model.graph;
 
 import jvmram.model.metrics.MetricType;
 
-public record GraphKey(MetricType type, long pid) {
+public record GraphKey(MetricType type, int pid) {
 }
